+++ carpenter-backend/src/main/java/com/js/carpenter/service/WoodService.java
@@ -3,11 +3,10 @@ package com.js.carpenter.service;
 import com.js.carpenter.model.RequiredItem;
 
 import java.util.List;
-import java.util.Map;
 
 public interface WoodService {
-    double getWoodLength();
-    boolean changeWoodLength(double length);
-    List<List<Double>> getCutting(List<RequiredItem> needs);
+    int getWoodLength();
+    boolean changeWoodLength(int length);
+    List<List<Integer>> getCutting(List<RequiredItem> needs);
 
 }
